requirements.txt
# ── Web Framework ──────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.0

# ── AI / LLM ───────────────────────────────────────
anthropic==0.28.0
transformers==4.44.0
torch==2.3.0+cu121 --extra-index-url https://download.pytorch.org/whl/cu121
accelerate==0.32.0
sentencepiece==0.2.0

# ── OCR (Windows specific) ─────────────────────────
Pillow==10.4.0
pdf2image==1.17.0
pytesseract==0.3.13
# NOTE: detectron2 is difficult on Windows. Use this pre-compiled version:
python-doctr==0.9.0             # Better alternative for Layout Analysis on Windows

# ── Embeddings & Vector Search ─────────────────────
sentence-transformers==3.1.0
pgvector==0.3.2
psycopg2-binary==2.9.9          # Must use -binary on Windows to avoid build errors

# ── De-identification ──────────────────────────────
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.6

# ── Data Processing ────────────────────────────────
pandas==2.2.2
numpy==1.26.4
datasets==2.21.0
kaggle==1.6.14
xmltodict==0.13.0
tqdm==4.66.4
python-dotenv==1.0.1

# ── Security & Testing ─────────────────────────────
cryptography==43.0.0
python-jose[cryptography]==3.3.0
pytest==8.3.2
httpx==0.27.0